Remove the user role, add bootstrap script and patches for pipenv
The user role handling is not necessary since our CI slaves
are going to be ephemeral hosts accessible via SSH key from the
housekeeping repository.

Introduce a bootstrap script that is going to prepare the environment
for provisioning (currently it just install ansible==2.2.0 via pip)

Add pacthes to work around the PIPE related deadlock in pipenv.