Support --ask-pass option in CLI
This uses getpass.getpass() to interactively retrieve a password
from an interactive prompt, and then sshpass to pass it along
to ssh via a custom file descriptor.

The associated test case similarly uses sshpass to supply the
test password to the LeApp CLI.

To avoid problems with the default per-tty credential caching
in sudo, the test case only runs when the tests are being
run as root (as is the case in CI).