Restore ability to run tests as non-root user
Due to recent changes in the way source & target VM lookup
is handled, the Cockpit plugin integration test now requires
that the VMs it uses be started by the root user (otherwise
it fails to find them).

While this isn't a problem for CI (where the tests run as
root), it causes problems when running the tests as a
normal user in a local `pipenv shell` session.

To fix that, the follow changes have been made:

- test scenarios can now indicate they need their VMs
  started as root, and if the tests are running as a
  normal user, the VMs will be destroyed and recreated
  by root
- test scenarios can now be marked with `@root_recommended`
  to indicate that they're really slow or otherwise
  problematic when run as a non-root user
- the Cockpit plugin integration test has been updated to
  use both of these features