Only use `sudo vagrant` when not root
By default, sudo calls don't inherit any enabled
Software Collections from the current environment.

This means `sudo vagrant` doesn't work by default
with the Vagrant SCL - the root session it creates
doesn't have the SCL enabled, so the command fails.

Rather than trying to enable the SCL in that case,
we instead just skip using sudo when the current
process is already running as root (as it is in CI).