Guard the __builtin__.open function before mock
When the __builtin__.open function is mocked inside a unit test,
test will fail because the environment cannot be initialised
correctly. Problem happens when the os.fdopenfcuntion is called
to create file object for the given stdin FD.

This is hackish workround, which store the original function object
during an import of the actor_definition module and ensure this
one is be temporary restored when os.fdopen is called.