Remove allow_null and missing from models fields
We do not see use-case to be able to distinguish whether value of
model's attribute is None or was not be set at all. The original
solution was confusing and there was not any actor writer that would
expect such non-python logic. To make everything clear, remove the
"missing" function and allow_null and simplify the logic just to
use the "required" value for fields.

So the current logic:
  - If required is set to True for a field, expect that field cannot
    contains the None value.
  - If required is set to False (default) for a field, the field can
    contains the None value.