upgrade: Auto reboot functionality
Previously reboots were always executed automatically, however this
behaviour isn't always wanted. Users and testers might need to do
things before the actual reboot.

This patch introduces a --reboot flag to the leapp upgrade
command line tool which will instruct the framework to perform
the reboot when it is requested. When the flag is not present
the upgrade will not performed and just a message will be presented
to the user asking them to reboot.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>