Fix usage of RawMessageDialog to print messages
Previously using RawMessageDialogs would cause a traceback when used.
This is because the dialog wanted to request something that wasn't
stored which would cause to return None. This has been fixed.

Additionally the RawMessageDialog overrides now the request_answers
implementation to perform the only necessary things for displaying
information to the user.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>