models: Support model references in fields.Model
Previously the Model field type, did not check whether or not it
received a reference as model to use, which resulted in some validation
error if the real model was passed to it. This PR fixes this problem
by trying to get the _resolved field from the model and defaults
to the actual model type that has been passed.

Fixes https://github.com/leapp-to/leapp/issues/374

Related-issue: https://github.com/leapp-to/leapp/issues/374
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>