actordefinition: Make stdin opening more robust (#365)
When executing leapp in a debugger it can happen that stdin can not
be opened with fdopen, in this case an OSError can occur. This patch
handles that scenario, but will imply that stdin might not be working
as expected in some cases. We won't be able to handle those cases but
this only happens within a debugger at this point. Therefore this
patch is good enough for now.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>