Introduce nullable field type
Currently we are using "missing" function as a placeholder
in cases where no value was specified in the particular field.
This might be confusing and can introduce issues when not
understood correctly. This patch is removing concept of the
missing function together with "allow_null" and "required"
parameters and switch to using a new nullable field type.
In this case every field which is not nullable has to be
specified. If we want to allow a field to be None we need to
wrap it around the nullable field.

E.g. fields.Nullable(fields.String())