Packaging: persistent TIMESTAMP (#384)
Previously, in case that between two calls of commands, where
the $(TIMESTAMP) have been included, have been processed in different
TIME (HH:MM), both commands result with different $(TIMESTAMP).

We want to proceed the make process with one persistant $(TIMESTAMP)
so we will avoid random errors related to the change of time. E.g.:
  cannot find prepared SRPM, because TIMESTAMP is different.
Problem is resolved in way that TIMESTAMP can be defined on the
cmdline. But to do not require it, special *_remake* target has been
added, which will call the make again with the TIMESTAMP value
in case the TIMESTAMP was not set.

Reproducer:
  1. Prepare Makefile like this:
    TIMESTAMP=`date +%Y%m%d%H%MZ -u`
    wait:
    	@echo $(TIMESTAMP)
    	@./wait_script.sh
    	@echo $(TIMESTAMP)
  2. call `make`