Do not fork actor execution with snactor fixture
Previously running an actor within a test caused the test to be executed
in a forked process and when it was executing the actor the actor was
also being executed in another forked process.

This patch is removing the latter fork from being done to allow easier
mocking and testing.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>