Add metapackage for dependencies
Currently we are not able to use python dependencies correctly to
ensure the RPM will not be removed during the upgrade of the system.
As a workround, put all such system dependencies into the leapp-deps
metapackage.

This metapackage is supposed to be built for the same system where
leapp should be installed. When different dependencies are needed
on target system (e.g. during the in-place upgrade) repository is
reponsible to deliver new meta package with such dependencies.

Leapp framework from this point will require specific capability:
  leapp-framework-dependencies

which has to be provided by such meta package. Any time the
dependencies are changed, the capability number has to be incremented
by one. So packages provided by leapp repository has to reflect such
change as well.