snactor: fix cryptic message without user repo config
If any snactor command is run before repos are
registered with snactor repo find command a
not much helping KeyError appears in error log,
because it isn't assumed that user repos may not
be configured at that point.
This patch addresses the issue by checking user repos
data presence and in case any missing links are found
raising RepositoryConfigurationError with a relevant
error message.