snactor: Workflow run --save-output
This implements a new feature for the `snactor workflow run` command,
that allows to store all messages from snactor workflow run to be
available for consumption by actor executed through `snactor run`.

Usage scenario:
---------------

One wants to write an actor that needs to consume from the
FactsCollection phase of the IPU workflow.

Before we introduced this parameter, one had to manually run every
actor, that is required to produce the messages the actor needs to
consume, manually with the --save-output parameter.

With the new feature one can now run:

$ snactor workflow run --save-output --until-phase FactsCollection ipu

And then can normally use `snactor run` to start testing the actor.

$ snactor run MyNewActor --print-output