testing: Support for retrieving snactor messages
In order to be able checking within unit tests whether or not, messages
have been saved to the database, it is necessary to retrieve the latest
snactor execution id. However since unit tests can run in multiple
repositories, it's necessary to be able to specify which database
connection to use.
This patch introduces some of the changes necessary for improving the
testability of `snactor workflow run --save-output` which will be
introduced in a follow up patch.