model.fields: Introduction of the JSON field type
The JSON field type, is a fields that allows to assign any value that is
encodable to JSON by the json library of the python standard library.
With the exception for None value as the value itself. This is to keep
consistency with the rest of the fields, that require the field to be
marked as nullable. None value nested within, however is not restricted
by this limitation.

Example usage:
```python
class JsonUsedModel(Model):
    topic = SomeTopic

    data = fields.JSON()
    nullable_data = fields.Nullable(fields.JSON())

m = JsonUsedModel()

m.data = {'valid': None} # valid value
m.data = None # invalid value for this field

m.nullable_data = None # valid
m.nullable_data = {'valid': None} # valid value
```

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>