stdlib: Functions to check the current config
Previously environment variables for debug and verbose mode had to be
manually checked by doing comparison to a string.

For example:

```python
if os.getenv('LEAPP_DEBUG', '0') == '1':
    do_something_debuggy()
if os.getenv('LEAPP_VERBOSE', '0') ==  '1':
    do_something_verbose()
```

This has been a constant source of mistakes, by forgetting to put the
numbers in quotes. In addition to this, it is very verbose and not very
convenient. Thus this change introduces two new functions in the
standard library that can be used to check those variables.

is_debug() will return True if LEAPP_DEBUG has been set to '1'
is_verbose() will return True if either LEAPP_DEBUG or LEAPP_VERBOSE are
set to '1'

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>