clicmd: Print help if a parent command is called
Previously when calling a command that has subcommand on the commandline
did not give the user any feedback on the usage.

As an example expected usage would be:
$ tool command subcommand

But the user would type:
$ tool command

The tool would silently and successfully exit.

With this change, the tool will now print the help and exit with exit
code 2.

Fixes #455

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>