checkpoints: Load checkpoints ordered by `id`
Previously checkpoints were loaded from the audit table in the database
and ordered by timestamp. However we ran into some problems were
timestamps were off due to some misconfiguration that caused the time
of the system being in the past after the reboot.
That resulted in leapp, trying to resume at a place that has been
already executed before.

This patch will change the ordering to use the row id, which is
sequential in SQLite.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>