Create a dynamic configuration phase
This commit adds a new "configuration" field into a workflow.
When a configuration model is added to the workflow and there is
an actor which produces it, the producing actor is run first as
part of the dynamically added phase and produces the config model
for consumption for all other actors, together with implementing
self.configuration api in every actor.

Example:

class FoobarWorkflow(Workflow):
    name = 'FoobarWorkflow'
    tag =  FoobarWorkflowTag
    short_name = 'foobar_workflow'
    description = '''No description has been provided ...'''
    configuration = FoobarConfig

    class FirstPhaseName(Phase):
       name = 'first_phase_name'
       filter = TagFilter(FirstPhaseTag)
       policies = Policies(Policies.Errors.FailPhase,
                           Policies.Retry.Phase)
       flags = Flags()

Then produce the configuration model using a "non-phase"
actor (see tags field):

class FoobarConfigProducer(Actor):
"""
No documentation has been provided ...
"""

    name = 'foobar_config_producer'
    consumes = ()
    produces = (FoobarConfig,)
    tags = (FoobarWorkflowTag,)

    def process(self):
        self.produce(FoobarConfig())

Co-authored-by: Vinzenz Feenstra <vfeenstr@redhat.com>