Fix: Undefined Model "ReportReference"
We are able to hit the issue only on some machines (not sure why just
on some machines) during the scan of the repository. The issue
happens during the loading workflow modules.

Modified snippet of the err:

    ...leapp.repository.common: Loading workflow modules
    ...leapp.repository.system_upgrade_el7toel8: Loading workflow modules

    Error: Undefined Model "ReportReference"

To fix that, import leapp.reporting inside the leapp.cli.upgrade
module to enure the Report model reference is loaded.