Make leapp show help message for proper subcommand
Before that patch leapp preupgrade --nosuchflag would
show help information for topmost leapp command, not
usage information for leapp preupgrade.
That is a clumsy hack to access help info for the
exact subcommand that has been called.
The original problem was with parser.parse_args()
referring to the parser for topmost leapp command and not
the subcommand during error handling.
Any ideas and proposals on a more elegant solution are
welcome.

Closes-Issue:#550