Log external command errors to audit
There's no reason why external command errors should not be logged in
audit.  In fact, having the error messages there may help understand the
context in which the command has failed.

Note that CalledProcessError is not actually logged as error; this is
something that should be probably dealt with by caller so it's not
always error per se.