Introduce skip_dialogs functionality
In order to make leapp preupgrade non-interactive by
default the workflow and actor being executed need to
be aware of the mode they are processed in (preupgrade or
upgrade workflow) and skip dialogs accordingly.

This is controlled by the skip_dialogs method variable which
if set to False by default and is propagated into every
workflow.run and actor initialization method.

The preupgrade cli command now sets skip_dialogs=True for
workflow.run which makes all dialogs present to be skipped and
answerfile with all options commented out to be generated while
the upgrade command keeps the old behavior allowing questions to
be asked interactively if no option has been chosen and saved in
answerfile.

Some basic workflow tests have been added.