packaging: improve dependencies and polish spec
Originally we had troubles with dependencies and release processes
as we have been strict about versions of leapp and repositories,
specified in the leapp rpm. As well, we would like to ensure that
currently we will install all rpms when leapp tool is installed.

So we are going to keep dependency on leapp-repository, but
without a version specification. We are going to left the
reponsiblity for version check on repositories itself.

As well, python?-leapp rpms will provide new leapp-framework
capability informing about incompatible changes or new features that
affects leapp repository (like changes in stdlib). This one will
be more flexible than official versioning and can be updated per
PR (see comment in the spec file for more details).

As well, I added several new macros in the SPEC, so there is no need
to update same values on several places anymore. Just update macros
in the top of the SPEC.