Make answerfile persistent between leapp preupgrade runs
This is a final touch to implementing the persistent answerfile
while still keeping the 1-1 mapping of answerfile-workflow dialogs
and allowing user to choose both leapp answer tool or editor of
choice for answerfile management.
The basic idea is refactor load_answerfile\save_answerfile into
load_answers\save_answers methods that apart from loading\saving
actual answerfile are also dumping the current state of answerstore
to the .userchoices file.
The leapp answer command will be used for answerfile management as
before.

https://github.com/oamg/leapp/issues/605