Documentations: changes in hierarchy + a new content
I realized it's hard to add new content specific to the el7toel8
leapp repository. As well, with new and new content, it's getting
harder and harder to find something in the menu as number of
items is growing.

For that purpose, here is proposed the new structure of documentation
that tries to reduce top level menu into categories that should cover
other content in future we would like to add.

TODO: it's still PoC. Another changes are expected yet.