Report inhibitors separately from errors
Previously there was no separation of errors and inhibitors in the
reports and the output of the leapp tool.
This patch introduces a new exception RequestStopAfterPhase that allows
to request the stop the execution of leapp gracefully after the current phase.
Additionally now all inhibiting reports are displayed in its own
section in the output.
As part of the output changes a bit of refactoring has been applied to
remove code duplication for 'pretty_block' use cases.

The leapp-framework version has been bumped to 1.1

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>