Fix handling of debug/verbose in {pre,}upgrade
Previously changes have been introduced that caused the debug/verbose
information not to be stored correctly to the database, which led to
different output behavior after reboot. (No debug/verbose output after
resuming the execution)

This patch changes the order when this data is processed so the
configuration is updated correctly.

JIRA-Taks: OAMG-3087
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>