Add --actor-config to snactor run
This patch introduces support for specifying a model which the actor
should consume for configuration data.
Using this feature requires to run an actor before which produces this
model and saves the output.

Additionally to facilitate the testing needs for this PR `snactor
messages add` has been added, which allows injecting messages from the
commandline.

Jira-Task: OAMG-1092

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>