Introduction of the @deprecated decorator
This patch introduces the @deprecated Decorator allowing it to be used
for Models and functions/methods anywhere in the code.
The code makes use of the python warnings library and extends the usage
around this by allowing to pass the information since when the entity is
deprecated and pass a reason for the deprecation (e.g. replacement by
something else)
On top of the @deprecated decorator there is now a @suppress_deprecation
decorator that allows to suppress deprecation warnings for given
entities. That allows to keep supporting deprecated items like Models
without triggering deprecation warnings.

Log deprecation warnings to the audit db

Additionally in this patch logging of deprecations to the audit database
is introduced.
It will create a new event type 'deprecation' and logs details about
location and place.

Start reporting deprecation usages

On top of it this patch adds deprecation usages to the reports.
It will receive the items from the audit database.
If an item is deprecated for more than 6 month the severity of
the report entry is set to HIGH otherwise it is set to MEDIUM.

Bumps framework_version to 1.3

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>