Allow caplog to see log records
Actors are executed in a different processes what
makes caplog blind to emitted messages.

Solution assumes adding a QueueHandler to the root logger
which will gather all the logs to the queue. Then we reemitt
log records from the queue in the main thread. This allows
caplog to see it