Improve documentation about deprecation process
The deprecation process is not clear as people do not understand
what to do when they want to deprecate a leapp entity. To help
people understand the deprecation process, let's add more text
to make it easier to accept the new process.

Some parts of the doc need addition improvements.
Issue: #663

Co-authored-by: ZhukovGreen <iam@zhukovgreen.pro>