Handle 'nested' suppress of deprecation
Imagine case:
```pseudocode
  (Ddep is deprecated entity)

  @suppress_deprecation(Ddep)
  def fA():
      create Ddep
      fB()

  @suppress_deprecation(Ddep)
  def fB():
      create Ddep
```

In such a case, apps crashed, because we store information about
suppressed deprecation into the set. After you leave the fB function
the Ddep entity is removed from the set and after end of fA function
there is no way to remove it again - so the app crashes. This commit
fixes the problem.