snactor.sanity-check: Add option to ignore models
Previously, if a Model was being consumed by actor before it was
produced,  `snactor sanity-check` reported a failure. This however
blocks some use case in leapp-repository now, where there is the
possibility to supply user data and it is being verified before in a
later stage it actually is used. This sanitization actor causes this
check to fail and would force an ugly workaround by having a dummy actor
pretending to produce the message earlier.

This patch will introduce two ways of handling this situation by
ignoring the models. Either each model can be specified as a commandline
option --ignore or a path to a file containing a list of such models can
be supplied via --ignorefile.
Any of those specified model names supplied via --ignore or the
ignore file will no longer be considered a failure.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>