Doc: add lists of the deprecated functionalites
Just document the stuff that is already deprecated. Added:
- a new document for the deprecated stuff in the el7toel8
  leapp repository
- a new section about deprecated stuff in leapp itself

The main deprecation document is pretty long now and the section
for the deprecated list in leapp is not good there. However I do not
see better way how to do it right now. The doc needs to be split
for sure in future, but this is not scope of this PR.