cli: Add command rerun
Per request this patch adds support for rerunning leapp from a certain
point. As for now it only supports the FirstBoot phase to be re-run.

In order to accomplish the re-run functionality leapp clones the context
of the last upgrade execution and removes all checkpoints and errors
from it.
Cloning means it will duplicate all the context aware data in the
database in order to allow a post mortem inspection of what has
happened.
So it will not blindly delete data but remove the errors and checkpoints
only from the cloned part of the data.

Additionally it allows to select only actors which have a certain tag
specified during the re-run. This will help users that want to re-run
their own set of actors only during the FirstBoot phase without having
to create a complicated workaround using snactor and somehow getting it
to use the data from the upgrade. This allows the actors that are re-run
to consume messages that have been generated during the upgrade before
this phase has been run.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>