models: Do not make references for private symbols
Previously there was a reference for Models created for every symbol
attempted to being imported from leapp.models, however pytest tries to
import `_pytestfixturefunction` which is not a model obviously.
To work around pytest causing failures because of that, private symbols
are being filtered from the model reference creation magic.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>