Add the _lookup_actors function into the repo manager
The RepositoryManager.lookup_actor function returns always just one
Actor or None. However, in case the actor of the same name is specified
in multiple repositories I would like to have an info about all these
actors. Currently I am not sure whether there is a mechanism to prevent
people having actors with identical name in multiple repositories when
leapp is running, but so far, tests do not react on the situation when
some actors of the same name exists. For now, creating this private
function to be able to run unit tests for actors of specific name
in all available repos.