Packaging: requires leapp-repository on RHEL 7 only
The leapp tool doesn't depend on leapp-repository anymore. However,
when no leapp repository is present on the system, there is nothing
to do usually when leapp is installed as all subcommands for leapp
should be provided by specific leapp repositorise.

I wanted to remove the dependency completly and keep it just in
downstream, however not sure whether that could be a good way.
So let's keep the requirement to provide the leapp-repository
at least as a capability for rpms that want to use leapp on RHEL 7.

People are going to be instructed to install e.g. `leapp-upgrade`
or `leapp-convert` to install everything what they need for the
action they want to do.

Additionally, use the `setup` macro instead of `autosetup`. We do not
expect to use any patches in upstream and the automatic application
was kind of fragile in downstream. So let's keep the `setup`
everywhere to be more synced.