Fix leapp prints nothing when running with no args on python3.6.
A change introduced in the stdlib argparse module accidentally caused
all subparsers to become optional instead of required as in python 2.7.
This change therefore causes leapp to print nothing and exit with error
code 0 when being executed without any arguments.

This commit marks the subparser action as required, so that the
python3.6 module recognizes this situation and prints the usage message.

To provide a meaningful message on python3.6 the subparser action was
edited to use 'command' metavar and the 'title' option was dropped.