Introduce the fast_lint make target.
Introduce fast_lint target enabling the developer to quickly lint only
files that have been modified on the current branch compared to master.

Also introduces the ENTER_VENV variable that expands to the command for
entering venv when not running as root or to empty string when running
as root, simplifying the recipes that require slightly different logic
when running as root or normal user.

This target contains the same recipe as the fast_lint target introduced
in leapp-repository.