Bump leapp-framework to 2.1
Regarding the recent changes, we provide new functionality regarding
get_*_path() functions - so the shared/common paths that do not
require execution of actors, can be used. This is useful e.g. when
we want to get specific files for the purpose of commands. Previous
use in actors is not affected by this change, however, use of these
functions outside of actors (e.g. in leapp commands) requires the
new implementation strictly.

Fixes: 742