Add python(abi) == 3.x deps for RHEL 8+ into the meta-pkg
On RHEL 8 it ould happen that people replace the original Python 3.6
by Python 3.8+. However, regarding the build & installation process,
we can support just ONE version of python per RPM - except the newer
version of python on the upgraded system, which is handled during
the upgrade by our workarounds. For that reason we put the abi
dependency back into the metapackage for RHEL 8 system only.

As well, added already python(abi) = 3.9 dependency for RHEL 9.