Fix of the fix for Python 3 depenencies
The previous fix with the python(abi) was wrong as it still could
lead into the situation when the system python is not installed,
because platform-python provides such ABI still. The correct fix
is to add dependency on python3 as such a dependency refers always
to the main python version for particular system despite the name
of the rpm package.

Bump leapp-framework-dependencies to 5