Fix store of answers on Python3.6 (DictProxy)
On Python3 the loaded answer is "sometimes" represented by DictProxy
object. In such a case the answer cannot be serialized into the
JSON object and storing into the audit db fails with crash.

"sometimes" ==> I have a suspicion the result is different in case
                the answer is loaded from the answerfile and
                userchoices file.

See: https://github.com/oamg/leapp-repository/pull/821

Note: This is not a proper fix. I am sending the commit from a train
      with unstable network.