testing: Start using GH Actions for unit tests (#719)
This patch applies all necessary changes to make tests run on
GitHub Actions.
As part of this patch we are transitioning like in leapp-repository
to ubi based containers running on podman. This will gives us the closest
resemblance to the target environments. Tests for fedora & others can be
eventually added as well. For now this patch focuses on our main targets
as the tests have not been working for some time.

Also as part of this patch some minor flake8 fixes have been applied,
and the answerstore test has been fixed.

Signed-off-by: Vinzenz Feenstra <vinzenz.feenstra@gmail.com>