Preserve the original inhibitor detection valid for now
As we substituted Tags and Flags with Groups recently, it has
impact on original detection of inhibitor inside leapp-repository
which is not compatible with the new reporting implementation.
As it could happen that new framework and old leapp-repository could
be installed on the system, the consequence in such a case is that
inhibitors will not be caught and user running:
  # leapp upgrade --reboot

could hit various serious issues. This change make sure we keep the
original functionality working, until we drop flags & tags
completely, providing enough time to everyone to adapt to new
changes.

Co-authored-by: Michal Hečko <michal.sk.com@gmail.com>
Co-authored-by: Ina Vasilevskaya <ivasilev@redhat.com>