Add Makefile target for building RPMs locally and in containers
Add `build` target to build locally and `build_container` to build
locally in a container. Building in containers has the benefit of being
able to build on all the different platforms e.g. both on  el7 and el8.

Run `make build` to build locally.

Run `make build_container` to build in a container. Note that
BUILD_CONTAINER environment variable must be set.

Run `clean_containers` to clean up container images used for building.

The containers are built from Containerfiles in `res/container-builds/`.
Since Containerfiles for all the Fedora versions would differ only in the `FROM`
statement, a generic Containerfile `Containerfile.fedora_generic` is
used and the `FROM` statement is inserted in the Makefile.

For el7 builds CentOS 7 container is used, because rpm-build package
is not available in ubi7 image.

Builds cannot be run in parallel because containers operate on the same
files, which would result in them conflicting.

Also update `.gitignore` to exclude build files from git.