Add reports summary to the output
Previously only the paths to reports were printed and as a result the
reports were easily missed.

Leapp now prints a list of HIGH and MEDIUM priority reports along with
a summary of number of reports with individual severities to make
reports more visible.

Also if there are any HIGH or MEDIUM severity reports, the block titles
are yellow.

The framework-version is bumped to 4.0, because the `report_info`
funtion now takes a context_id parameter (to be able to fetch the
reports), so the new API is incompatible.