Improve handling of a remediation with non-ascii data
When a remediation command is printed that manipulates
non-ascii but totally valid utf-8 strings, leapp should
not give out any tracebacks upon report generation.
The idea of a fix is to make repr implementation of Remediation
aware of utf-8 encoding.
The root of the problem is that a mixture of native text / encoded
string can appear in a report message and that caused report generating
code go crazy while creating actual log files.
Unit test has been added as well.

Related-to: OAMG-8629