Add an ERROR report group
Currently there is no straightforward way to identify whether a report
is an ordinary report or an error report (report generated from an
error).

An `ERROR` group is introduced in `leapp.reporting.Groups.ERROR` which
is used to group error reports. This is utilized in the leapp output to
show the number of error reports in the summary. Also error reports in
`leapp-report.txt` are now followed by a "(error)" string similar to
inhibitors.

For now ERROR isn't in flags nor tags in the leapp-report.json in
report schema 1.1.0. This because integration with plugins might break,
due to the enumeration of possible flags in the schema. This will be
adjusted later. In report schema 1.2.0 the ERROR group is listed in
groups as expected.

Jira: OAMG-9532