Add an ERROR report group
Currently there is no straightforward way to identify whether a report
is an ordinary report or an error report (report generated from an
error).

An `ERROR` group is introduced in `leapp.reporting.Groups.ERROR` which
is used to group error reports. This is utilized in the leapp output to
show the number of error reports in the summary. Also error reports in
`leapp-report.txt` are now followed by a "(error)" string similar to
inhibitors.

Jira: OAMG-9532